connected redux form with current form
