+++ package.json
@@ -37,6 +37,7 @@
     "@types/node": "^10.12.18",
     "@types/react": "^16.7.20",
     "@types/react-dom": "^16.0.11",
+    "@types/react-redux": "^7.0.1",
     "@types/redux-form": "^7.4.16",
     "@types/styled-components": "^4.1.6",
     "connected-react-router": "^6.2.2",
