+++ src/ui/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Helper } from "./Helper";
 export { default as Input } from "./Input";
+export { default as Navbar } from "./Navbar";
