+++ src/ui/containers/Home.ts
@@ -1,0 +1,11 @@
+import { connect } from "react-redux";
+import { reduxForm } from "redux-form";
+
+import { Home } from "../screens";
+
+const DecoratedHomeForm = reduxForm({
+  form: "CronConfigForm",
+  destroyOnUnmount: false
+})(Home);
+
+export const HomeContainer = connect()(DecoratedHomeForm);
