+++ src/ui/screens/About/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const About = () => (
+  <div>
+    <h2>About</h2>
+  </div>
+);
+
+export default About;
