+++ src/ui/screens/index.ts
@@ -1,0 +1,2 @@
+export { default as About } from "./About";
+export { default as Home } from "./Home";
